% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatCamelCase.R
\name{formatCamelCase}
\alias{formatCamelCase}
\title{format a give string to camel case (e.g., \emph{Abc_Efg})}
\usage{
formatCamelCase(x, sep = "_")
}
\arguments{
\item{x}{input string}

\item{sep}{separator. Default \emph{_}.}
}
\value{
the string in camel format
}
\description{
format a give string to camel case (e.g., \emph{Abc_Efg})
}
