% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3Lollipop.R
\name{g3Lollipop-shiny}
\alias{g3Lollipop-shiny}
\alias{g3LollipopOutput}
\alias{renderG3Lollipop}
\title{Shiny bindings for g3Lollipop}
\usage{
g3LollipopOutput(outputId, width = "100\%", height = "520px")

renderG3Lollipop(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a g3-lollipop}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
No value returned. It is the binding which enables interactive functions
  within Shiny applications and Rmd documents.
}
\description{
Output and render functions for using g3viz lollipop diagram within Shiny
applications and interactive Rmd documents.
}
