% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3LollipopThemeOptions.R
\name{g3Lollipop.theme.options}
\alias{g3Lollipop.theme.options}
\title{G3Lollipop chart themes}
\usage{
g3Lollipop.theme.options(theme.name = "default", title.text = "",
  y.axis.label = "# of mutations", legend.title = NA)
}
\arguments{
\item{theme.name}{theme name.  Default \emph{default}}

\item{title.text}{title of chart. Default is empty.}

\item{y.axis.label}{Y-axis label text. Default \emph{"# of mutations"}.}

\item{legend.title}{legend title.  If \code{NA}, \emph{factor.col} in \code{\link{g3Lollipop}} is used.
Default is \code{NA}.}
}
\value{
a list with g3Lollipop plot options

a list with g3Lollipop plot options
}
\description{
G3Lollipop chart themes
}
