% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3LollipopTheme.R
\name{g3Lollipop.theme}
\alias{g3Lollipop.theme}
\title{G3Lollipop chart options of built-in themes.}
\usage{
g3Lollipop.theme(
  theme.name = "default",
  title.text = "",
  y.axis.label = "# of mutations",
  legend.title = NA
)
}
\arguments{
\item{theme.name}{theme name, including \emph{default}, \emph{cbioportal},
\emph{nature}, \emph{nature2}, \emph{dark}, \emph{blue}, \emph{ggplot2}, and \emph{simple}. Default
\emph{default}.}

\item{title.text}{title of chart. Default is empty.}

\item{y.axis.label}{Y-axis label text. Default \emph{"# of mutations"}.}

\item{legend.title}{legend title.  If \code{NA}, \emph{factor.col} in
\code{\link{g3Lollipop}} is used. Default is \code{NA}.}
}
\value{
a list with g3Lollipop plot options
}
\description{
G3Lollipop chart options of built-in themes.
}
