% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessMAFColumn.R
\name{guessMAFColumnName}
\alias{guessMAFColumnName}
\title{Guess column name for MAF file}
\usage{
guessMAFColumnName(maf.df, alt.column.names)
}
\arguments{
\item{maf.df}{MAF data frame}

\item{alt.column.names}{a vector of alternative column names}
}
\value{
if hit one alternative column name, return the name; otherwise, return \code{NA}
}
\description{
Guess column name for MAF file
}
