% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgnc2pfam.df}
\alias{hgnc2pfam.df}
\title{Mapping table between gene.symbol, uniprot.id, and pfam}
\format{A data frame with columns of symbol, uniprot, length, start, end, hmm.acc, hmm.name, type}
\source{
Pfam (v31.0) and UniProt
}
\usage{
hgnc2pfam.df
}
\description{
A dataset containing the mapping table between Hugo symbol, UniProt ID, and Pfam ACC.
}
\keyword{datasets}
