% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMutationTypeToMutationClass.R
\name{mapMutationTypeToMutationClass}
\alias{mapMutationTypeToMutationClass}
\title{Map from mutation type to mutation class}
\usage{
mapMutationTypeToMutationClass(mutation.type.vec,
  mutation.type.to.class.df = NULL)
}
\arguments{
\item{mutation.type.vec}{a vector of mutation type information}

\item{mutation.tytpe.to.class.df}{Data frame mapping from mutation tytpe (header \emph{Mutation_Type})
to mutation class (header \emph{Mutation_Class}).
Default is \code{\link{getDefaultMutationMappingTable}}.}
}
\value{
a vector of mapped mutation class information
}
\description{
Map from mutation type to mutation class. Refer to \code{\link{getDefaultMutationMappingTable}}
for more details.
}
