% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutation.table.df}
\alias{mutation.table.df}
\title{Default mapping table between mutation type (aka, variant classification) to mutation class}
\format{A data frame with three columns, \emph{Mutation_Type}, \emph{Mutation_Class}, and \emph{Short_Name}.}
\usage{
mutation.table.df
}
\description{
A dataset containing the mapping table between genomic mutation type (aka, variant classification) to mutation class
}
\keyword{datasets}
