% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseProteinChange.R
\name{parseProteinChange}
\alias{parseProteinChange}
\title{Extract \emph{amino_acid_position} from \emph{Protein_Change}}
\usage{
parseProteinChange(protein.change.vec, mutation.class.vec)
}
\description{
Parse \emph{amino_acid_position} according to HGVSp format. \cr
 For example, \emph{p.Q16Rfs*28}, amino-acid position is 16.
  See \url{http://varnomen.hgvs.org/recommendations/protein/}. \cr
  See \url{https://www.hgvs.org/mutnomen/recs-prot.html}. \cr
}
