% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3Lollipop.R
\name{g3Lollipop}
\alias{g3Lollipop}
\title{G3Lollipop diagram for mutation data}
\usage{
g3Lollipop(mutation.dat, gene.symbol, uniprot.id = NA,
  gene.symbol.col = "Hugo_Symbol", aa.pos.col = "AA_Position",
  protein.change.col = c("Protein_Change", "HGVSp_Short"),
  factor.col = "Mutation_Class", plot.options = list())
}
\arguments{
\item{mutation.dat}{Mutation data frame.}

\item{gene.symbol}{HGNC gene symbol.}

\item{uniprot.id}{UniProt ID, in case that gene symbol maps to multiple UniProt entries.}

\item{gene.symbol.col}{Column name of Hugo gene symbols (e.g., TP53). Default \emph{Hugo_Symbol}.}

\item{aa.pos.col}{Column name of the parsed amino-acid change position. Default \emph{AA_Position}.}

\item{protein.change.col}{Column name of protein change information (e.g., p.K960R, G658S, L14Sfs*15).
Default is a list of \emph{Protein_Change}, \emph{HGVSp_Short}.}

\item{factor.col}{column of classes in the plot legend. IF \code{NA}, use parsed \emph{Mutation_Class} column,
otherwise, use specified.  Default \code{NA}.}

\item{plot.options}{options of lollipop plot in list format}
}
\description{
Generate G3Lollipop chart from the given mutation data.
}
