% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutationsFromCbioportal.R
\name{getMutationsFromCbioportal}
\alias{getMutationsFromCbioportal}
\title{Query cancer genomic mutation data from cBioPortal}
\usage{
getMutationsFromCbioportal(study.id, gene.symbol, output.file = NA,
  mutation.type.to.class.df = NA)
}
\arguments{
\item{study.id}{cbioprotal study ID}

\item{gene.symbol}{HGNC gene symbol.}

\item{output.file}{if specified, output to a file in \emph{CSV} format. Default is \code{NA}.}

\item{mutation.type.to.class.df}{mapping table from mutation type to class.
\code{\link{getDefaultMutationMappingTable}} for details.
Default \code{NA}, indicating to use \code{\link{getDefaultMutationMappingTable}}.}
}
\value{
a data frame with columns
\itemize{
\item \emph{Hugo_Symbol} --- Hugo gene symbol
\item \emph{Protein_Change} --- Protein change information (cBioprotal uses \emph{HGVSp} format)
\item \emph{Sample_ID} --- Sample ID
\item \emph{Mutation_Type} --- mutation type, i.e.,
\item \emph{Chromosome} --- chromosome
\item \emph{Start_Position} --- start position
\item \emph{End_Position} --- end position
\item \emph{Reference_Allele} --- reference allele
\item \emph{Variant_Allele} --- variant allele
\item \emph{Mutation_Class} --- mutation class (e.g., Truncating/Misense/Inframe/Other)
\item \emph{AA_Position} --- amino-acid postion of the variant; if the variant is not in protein-conding region, \code{NA}
}
}
\description{
Retrieve and parse mutation data from cBioPortal
by the given cBioPortal cancer study ID and the gene symbol.
}
\examples{
# list all cbioportal studies
library(cgdsr)
cgds <- CGDS("http://www.cbioportal.org/public-portal/")
all.studies <- getCancerStudies(cgds)

# pick a "caner_study_id" (contain a mutation data set)
# pick a primary HGNC gene symbol to query
mutation.dat <- getMutationsFromCbioportal("msk_impact_2017", "TP53")
mutation.dat <- getMutationsFromCbioportal("all_stjude_2016", "TP53")
}
