% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutationsFromCbioportal.R
\name{getMutationsFromCbioportal}
\alias{getMutationsFromCbioportal}
\title{Query cancer genomic mutation data from cBioPortal}
\usage{
getMutationsFromCbioportal(
  study.id,
  gene.symbol,
  output.file = NA,
  mutation.type.to.class.df = NA
)
}
\arguments{
\item{study.id}{cbioprotal study ID}

\item{gene.symbol}{HGNC gene symbol.}

\item{output.file}{if specified, output to a file in \emph{CSV} format.
Default is \code{NA}.}

\item{mutation.type.to.class.df}{mapping table from mutation type to class.
See \code{\link{mapMutationTypeToMutationClass}} for details. Default
\code{NA}, which indicates to use default mappings.}
}
\value{
a data frame with columns
  \describe{
      \item{Hugo_Symbol}{Hugo gene symbol}
      \item{Protein_Change}{Protein change information (cBioportal uses \emph{HGVSp} format)}
      \item{Sample_ID}{Sample ID}
      \item{Mutation_Type}{mutation type, aka, variant classification.}
      \item{Chromosome}{chromosome}
      \item{Start_Position}{start position}
      \item{End_Position}{end position}
      \item{Reference_Allele}{reference allele}
      \item{Variant_Allele}{variant allele}
      \item{Mutation_Class}{mutation class (e.g., Truncating/Missense/Inframe/Other)}
      \item{AA_Position}{amino-acid position of the variant; if the variant is not in
      protein-coding region, \code{NA}}
      }
}
\description{
Retrieve and parse mutation data from cBioPortal by the given
  cBioPortal cancer study ID and the gene symbol.
}
\examples{
\dontrun{
# Usage:
# cBioPortalData has officially replaced the defunct cgdsr.
# Search online for cgdsrMigration.html if interested.
library(cBioPortalData)
cbio <- cBioPortal()

# list all studies of cBioPortal
all.studies <- getStudies(cbio, buildReport = FALSE)

# First, select a cancer study that contains mutation data set ("caner_study_id")
# then, query genomic mutation data using a HGNC gene symbol,
# for example
mutation.dat <- getMutationsFromCbioportal("msk_impact_2017", "TP53")
mutation.dat <- getMutationsFromCbioportal("all_stjude_2016", "TP53")
}
}
