% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMutationTypeToMutationClass.R
\name{mapMutationTypeToMutationClass}
\alias{mapMutationTypeToMutationClass}
\title{Map from mutation type to mutation class}
\usage{
mapMutationTypeToMutationClass(mutation.type.vec,
  mutation.type.to.class.df = NA)
}
\arguments{
\item{mutation.type.vec}{a vector of mutation type information}

\item{mutation.type.to.class.df}{A mapping table from mutation type (header \emph{Mutation_Type})
to mutation class (header \emph{Mutation_Class}).
Default \code{NA} indicates to use default mapping table; see \code{\link{getDefaultMutationMappingTable}}.}
}
\value{
a vector of mapped mutation class information
}
\description{
Map from mutation type to mutation class. Refer to \code{\link{getDefaultMutationMappingTable}}
for more details.
}
