% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseProteinChange.R
\name{parseProteinChange}
\alias{parseProteinChange}
\title{Extract \emph{amino_acid_position} from \emph{Protein_Change}}
\usage{
parseProteinChange(protein.change.vec, mutation.class.vec)
}
\arguments{
\item{protein.change.vec}{a vector of strings with protein change information, usually in HGVSp_short format.}

\item{mutation.class.vec}{a vector of strings with mutation class (or so-called variant classification) information.}
}
\value{
a vector of parsed amino-acid position
}
\description{
Parse \emph{amino_acid_position} according to HGVSp_short format. \cr
 For example, \emph{p.Q16Rfs*28}, amino-acid position is 16.
  See \url{http://varnomen.hgvs.org/recommendations/protein/} or \url{https://www.hgvs.org/mutnomen/recs-prot.html}. \cr
}
